Added lenient disconnect for slaves
Makes it possible to set slaves temporarily offline leniently.
This means that they are not set as offline until the downstream
is completed. However, if another slave is able to take over
the downstream builds, the slave in lenient offline mode can be
taken down directly after the current build is finished.

Change-Id: I88eeb529c9cccdc3542217ec75b5a973956d7fc3